pyproject.toml
[tool.poetry]
name = "customer-outflow"
version = "0.1.0"
description = "Customer churn forecasting for Telecommunication company"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]



[tool.poetry.dependencies]
python = ">=3.10"
pandas = "2.2.3"
numpy = "^1.24.3"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
scikit-learn = "1.6.0"
scipy = "1.14.1"
dash = "^2.11.0"
faker = "^33.1.0"
nbformat = "^5.10.4"
pdoc = "^15.0.1"
plotly-express = "^0.4.1"




[tool.poetry.dev-dependencies]
pytest = "^7.2"
black = "^23.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
